pgtype Int2, Int4, and Int8 fields include bit size
e.g. Instead of Int it is Int64. This matches the pattern set by the
database/sql types.

+++ pgtype/builtin_wrappers.go
@@ -18,19 +18,19 @@ func (w *int8Wrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *int8")
 	}
 
-	if v.Int < math.MinInt8 {
-		return fmt.Errorf("%d is less than minimum value for int8", v.Int)
+	if v.Int64 < math.MinInt8 {
+		return fmt.Errorf("%d is less than minimum value for int8", v.Int64)
 	}
-	if v.Int > math.MaxInt8 {
-		return fmt.Errorf("%d is greater than maximum value for int8", v.Int)
+	if v.Int64 > math.MaxInt8 {
+		return fmt.Errorf("%d is greater than maximum value for int8", v.Int64)
 	}
-	*w = int8Wrapper(v.Int)
+	*w = int8Wrapper(v.Int64)
 
 	return nil
 }
 
 func (w int8Wrapper) Int64Value() (Int8, error) {
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 type int16Wrapper int16
@@ -42,19 +42,19 @@ func (w *int16Wrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *int16")
 	}
 
-	if v.Int < math.MinInt16 {
-		return fmt.Errorf("%d is less than minimum value for int16", v.Int)
+	if v.Int64 < math.MinInt16 {
+		return fmt.Errorf("%d is less than minimum value for int16", v.Int64)
 	}
-	if v.Int > math.MaxInt16 {
-		return fmt.Errorf("%d is greater than maximum value for int16", v.Int)
+	if v.Int64 > math.MaxInt16 {
+		return fmt.Errorf("%d is greater than maximum value for int16", v.Int64)
 	}
-	*w = int16Wrapper(v.Int)
+	*w = int16Wrapper(v.Int64)
 
 	return nil
 }
 
 func (w int16Wrapper) Int64Value() (Int8, error) {
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 type int32Wrapper int32
@@ -66,19 +66,19 @@ func (w *int32Wrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *int32")
 	}
 
-	if v.Int < math.MinInt32 {
-		return fmt.Errorf("%d is less than minimum value for int32", v.Int)
+	if v.Int64 < math.MinInt32 {
+		return fmt.Errorf("%d is less than minimum value for int32", v.Int64)
 	}
-	if v.Int > math.MaxInt32 {
-		return fmt.Errorf("%d is greater than maximum value for int32", v.Int)
+	if v.Int64 > math.MaxInt32 {
+		return fmt.Errorf("%d is greater than maximum value for int32", v.Int64)
 	}
-	*w = int32Wrapper(v.Int)
+	*w = int32Wrapper(v.Int64)
 
 	return nil
 }
 
 func (w int32Wrapper) Int64Value() (Int8, error) {
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 type int64Wrapper int64
@@ -90,13 +90,13 @@ func (w *int64Wrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *int64")
 	}
 
-	*w = int64Wrapper(v.Int)
+	*w = int64Wrapper(v.Int64)
 
 	return nil
 }
 
 func (w int64Wrapper) Int64Value() (Int8, error) {
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 type intWrapper int
@@ -108,20 +108,20 @@ func (w *intWrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *int")
 	}
 
-	if v.Int < math.MinInt {
-		return fmt.Errorf("%d is less than minimum value for int", v.Int)
+	if v.Int64 < math.MinInt {
+		return fmt.Errorf("%d is less than minimum value for int", v.Int64)
 	}
-	if v.Int > math.MaxInt {
-		return fmt.Errorf("%d is greater than maximum value for int", v.Int)
+	if v.Int64 > math.MaxInt {
+		return fmt.Errorf("%d is greater than maximum value for int", v.Int64)
 	}
 
-	*w = intWrapper(v.Int)
+	*w = intWrapper(v.Int64)
 
 	return nil
 }
 
 func (w intWrapper) Int64Value() (Int8, error) {
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 type uint8Wrapper uint8
@@ -133,19 +133,19 @@ func (w *uint8Wrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *uint8")
 	}
 
-	if v.Int < 0 {
-		return fmt.Errorf("%d is less than minimum value for uint8", v.Int)
+	if v.Int64 < 0 {
+		return fmt.Errorf("%d is less than minimum value for uint8", v.Int64)
 	}
-	if v.Int > math.MaxUint8 {
-		return fmt.Errorf("%d is greater than maximum value for uint8", v.Int)
+	if v.Int64 > math.MaxUint8 {
+		return fmt.Errorf("%d is greater than maximum value for uint8", v.Int64)
 	}
-	*w = uint8Wrapper(v.Int)
+	*w = uint8Wrapper(v.Int64)
 
 	return nil
 }
 
 func (w uint8Wrapper) Int64Value() (Int8, error) {
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 type uint16Wrapper uint16
@@ -157,19 +157,19 @@ func (w *uint16Wrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *uint16")
 	}
 
-	if v.Int < 0 {
-		return fmt.Errorf("%d is less than minimum value for uint16", v.Int)
+	if v.Int64 < 0 {
+		return fmt.Errorf("%d is less than minimum value for uint16", v.Int64)
 	}
-	if v.Int > math.MaxUint16 {
-		return fmt.Errorf("%d is greater than maximum value for uint16", v.Int)
+	if v.Int64 > math.MaxUint16 {
+		return fmt.Errorf("%d is greater than maximum value for uint16", v.Int64)
 	}
-	*w = uint16Wrapper(v.Int)
+	*w = uint16Wrapper(v.Int64)
 
 	return nil
 }
 
 func (w uint16Wrapper) Int64Value() (Int8, error) {
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 type uint32Wrapper uint32
@@ -181,19 +181,19 @@ func (w *uint32Wrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *uint32")
 	}
 
-	if v.Int < 0 {
-		return fmt.Errorf("%d is less than minimum value for uint32", v.Int)
+	if v.Int64 < 0 {
+		return fmt.Errorf("%d is less than minimum value for uint32", v.Int64)
 	}
-	if v.Int > math.MaxUint32 {
-		return fmt.Errorf("%d is greater than maximum value for uint32", v.Int)
+	if v.Int64 > math.MaxUint32 {
+		return fmt.Errorf("%d is greater than maximum value for uint32", v.Int64)
 	}
-	*w = uint32Wrapper(v.Int)
+	*w = uint32Wrapper(v.Int64)
 
 	return nil
 }
 
 func (w uint32Wrapper) Int64Value() (Int8, error) {
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 type uint64Wrapper uint64
@@ -205,11 +205,11 @@ func (w *uint64Wrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *uint64")
 	}
 
-	if v.Int < 0 {
-		return fmt.Errorf("%d is less than minimum value for uint64", v.Int)
+	if v.Int64 < 0 {
+		return fmt.Errorf("%d is less than minimum value for uint64", v.Int64)
 	}
 
-	*w = uint64Wrapper(v.Int)
+	*w = uint64Wrapper(v.Int64)
 
 	return nil
 }
@@ -219,7 +219,7 @@ func (w uint64Wrapper) Int64Value() (Int8, error) {
 		return Int8{}, fmt.Errorf("%d is greater than maximum value for int64", w)
 	}
 
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 type uintWrapper uint
@@ -231,15 +231,15 @@ func (w *uintWrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *uint64")
 	}
 
-	if v.Int < 0 {
-		return fmt.Errorf("%d is less than minimum value for uint64", v.Int)
+	if v.Int64 < 0 {
+		return fmt.Errorf("%d is less than minimum value for uint64", v.Int64)
 	}
 
-	if uint64(v.Int) > math.MaxUint {
-		return fmt.Errorf("%d is greater than maximum value for uint", v.Int)
+	if uint64(v.Int64) > math.MaxUint {
+		return fmt.Errorf("%d is greater than maximum value for uint", v.Int64)
 	}
 
-	*w = uintWrapper(v.Int)
+	*w = uintWrapper(v.Int64)
 
 	return nil
 }
@@ -249,7 +249,7 @@ func (w uintWrapper) Int64Value() (Int8, error) {
 		return Int8{}, fmt.Errorf("%d is greater than maximum value for int64", w)
 	}
 
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 type float32Wrapper float32
@@ -261,7 +261,7 @@ func (w *float32Wrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *float32")
 	}
 
-	*w = float32Wrapper(v.Int)
+	*w = float32Wrapper(v.Int64)
 
 	return nil
 }
@@ -271,7 +271,7 @@ func (w float32Wrapper) Int64Value() (Int8, error) {
 		return Int8{}, fmt.Errorf("%f is greater than maximum value for int64", w)
 	}
 
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 func (w *float32Wrapper) ScanFloat64(v Float8) error {
@@ -297,7 +297,7 @@ func (w *float64Wrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *float64")
 	}
 
-	*w = float64Wrapper(v.Int)
+	*w = float64Wrapper(v.Int64)
 
 	return nil
 }
@@ -307,7 +307,7 @@ func (w float64Wrapper) Int64Value() (Int8, error) {
 		return Int8{}, fmt.Errorf("%f is greater than maximum value for int64", w)
 	}
 
-	return Int8{Int: int64(w), Valid: true}, nil
+	return Int8{Int64: int64(w), Valid: true}, nil
 }
 
 func (w *float64Wrapper) ScanFloat64(v Float8) error {
@@ -346,7 +346,7 @@ func (w *stringWrapper) ScanInt64(v Int8) error {
 		return fmt.Errorf("cannot scan NULL into *string")
 	}
 
-	*w = stringWrapper(strconv.FormatInt(v.Int, 10))
+	*w = stringWrapper(strconv.FormatInt(v.Int64, 10))
 
 	return nil
 }
@@ -357,7 +357,7 @@ func (w stringWrapper) Int64Value() (Int8, error) {
 		return Int8{}, err
 	}
 
-	return Int8{Int: int64(num), Valid: true}, nil
+	return Int8{Int64: int64(num), Valid: true}, nil
 }
 
 type timeWrapper time.Time
